+++ tests/unit/test_client.py
@@ -15,23 +15,23 @@ def client(requests_mock):
     yield client
 
 
-class TestGetInfo(object):
-    def test_get_one(self, requests_mock, client):
+class TestGetData(object):
+    def test_get_one_data(self, requests_mock, client):
         data = {"id": 1, "name": "image"}
         requests_mock.get(
             f"{client._api_url}images/1/", json=data,
         )
-        result = client.get_info("images", 1)
+        result = client.get_one_data("images", 1)
 
         assert result == data
 
-    def test_get_list(self, requests_mock, client):
+    def test_get_list_data(self, requests_mock, client):
         data = [{"id": 1, "name": "image"}, {"id": 2, "name": "image"}]
         requests_mock.get(
             f"{client._api_url}images/",
             json={"next": None, "previous": None, "maxPage": 1, "data": data},
         )
-        result = client.get_info("images")
+        result = client.get_list_data("images")
 
         assert result == data
 
@@ -140,11 +140,12 @@ class TestAnalysisKeypoint(object):
     def test_キーポイント解析を開始できること(self, requests_mock, client):
         keypoint_id = 111
         expected_analysis_id = 222
+        rule = []
         requests_mock.post(
             f"{client._api_url}analyses/", json={"id": expected_analysis_id}
         )
 
-        analysis_id = client.analyze_keypoint(keypoint_id)
+        analysis_id = client.analyze_keypoint(keypoint_id, rule)
 
         assert analysis_id == expected_analysis_id